% Auto-generated: do not edit by hand
\name{dashSwiper}

\alias{dashSwiper}

\title{DashSwiper component}

\description{

}

\usage{
dashSwiper(id=NULL, autoplay=NULL, className=NULL, loop=NULL,
navigation=NULL, nextButton=NULL, pagination=NULL,
prevButton=NULL, shader=NULL, slides=NULL, speed=NULL,
swiperOptions=NULL)
}

\arguments{
\item{id}{Character. }

\item{autoplay}{Lists containing elements 'delay', 'disableoninteraction'.
those elements have the following types:
  - delay (numeric; optional)
  - disableoninteraction (logical; optional). }

\item{className}{Character. }

\item{loop}{Logical. }

\item{navigation}{Lists containing elements 'prevel', 'nextel'.
those elements have the following types:
  - prevel (character; optional)
  - nextel (character; optional). }

\item{nextButton}{Logical. }

\item{pagination}{Lists containing elements 'el', 'clickable'.
those elements have the following types:
  - el (character; optional)
  - clickable (logical; optional). }

\item{prevButton}{Logical. }

\item{shader}{Character. }

\item{slides}{List of lists containing elements 'src', 'alt', 'title', 'link'.
those elements have the following types:
  - src (character; required)
  - alt (character; optional)
  - title (character; optional)
  - link (character; optional)s. }

\item{speed}{Numeric. }

\item{swiperOptions}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

